[add]削除処理追加
