+++ src/components/network/Network.tsx
@@ -42,6 +42,7 @@ import PlaylistAddIcon from "@mui/icons-material/PlaylistAdd";
 import SimilaritySettings from "./similaritySettings/SimilaritySettings";
 import { fetcher } from "../common/Fetcher";
 import useSWR from "swr";
+import SearchGames from "../SearchGames/SearchGames";
 
 
 type Props = {
@@ -75,15 +76,6 @@ const Network = (props: Props) => {
 
   const { tourRun, setTourRun } = useTour();
 
-  // Sidebarが開いているかどうかを管理する状態
-  const [isSidebarOpen, setIsSidebarOpen] = useState<boolean>(false);
-
-  // 検索関連の状態
-  const [searchQuery, setSearchQuery] = useState<string>("");
-  const [filteredSteamList, setFilteredSteamList] = useState<SteamListType[]>(
-    []
-  );
-  const [isFocused, setIsFocused] = useState<boolean>(false);
   const [userAddedGames, setUserAddedGames] = useState<string[]>([]);
 
   const initialNodes = async (
@@ -132,7 +124,6 @@ const Network = (props: Props) => {
     if (selectedIndex !== -1 && nodes[selectedIndex]) {
       setCenterX((nodes[selectedIndex].x ?? 0) - 150);
       setCenterY((nodes[selectedIndex].y ?? 0) + 100);
-      // ノードが選択されたら GameSearchPanel を開く
       setIsGameSearchOpen(true);
     }
   }, [selectedIndex]);
@@ -140,7 +131,6 @@ const Network = (props: Props) => {
   const togglePanel = (panelName: string) => {
     setOpenPanel((prevPanel) => {
       const newPanel = prevPanel === panelName ? null : panelName;
-      setIsSidebarOpen(newPanel !== null); // Sidebarが開いているかどうかを更新
       return newPanel;
     });
     setTourRun(false);
@@ -157,39 +147,6 @@ const Network = (props: Props) => {
     });
   };
 
-  // 検索ロジック
-  useEffect(() => {
-    const lowerCaseQuery = searchQuery.toLowerCase();
-
-    const allSteamList: SteamListType[] = nodes.map(
-      (node: NodeType): SteamListType => {
-        return {
-          steamGameId: node.steamGameId,
-          title: node.title,
-          index: node.index,
-        };
-      }
-    );
-
-    allSteamList.push(
-      ...steamListData.filter(
-        (item1: SteamListType) =>
-          !allSteamList.find(
-            (item2: SteamListType) => item2.steamGameId === item1.steamGameId
-          )
-      )
-    );
-
-    const filteredSteam = allSteamList
-      .filter(
-        (game) =>
-          game.title.toLowerCase().includes(lowerCaseQuery) && game.steamGameId
-      )
-      .slice(0, 20);
-
-    setFilteredSteamList(filteredSteam);
-  }, [searchQuery, userAddedGames, nodes, steamListData]);
-
   // ユーザーが追加したゲームを取得
   useEffect(() => {
     (async () => {
@@ -198,49 +155,6 @@ const Network = (props: Props) => {
     })();
   }, []);
 
-  // ゲームを追加する処理
-  const handleSearchClick = (steamGameId: string) => {
-    if (
-      !userAddedGames.includes(steamGameId) &&
-      !nodes.some((node) => node.steamGameId === steamGameId)
-    ) {
-      const newUserAddedGames = [...userAddedGames, steamGameId];
-      setUserAddedGames(newUserAddedGames);
-      (async () => {
-        await changeGameIdData(newUserAddedGames);
-        setSearchQuery("");
-        setIsNetworkLoading(true);
-      })();
-    }
-  };
-
-  // ゲームを削除する処理
-  const handleGameDelete = (steamGameId: string) => {
-    const newUserAddedGames = userAddedGames.filter(
-      (gameId: string) => gameId !== steamGameId
-    );
-    setUserAddedGames(newUserAddedGames);
-    (async () => {
-      await changeGameIdData(newUserAddedGames);
-      setSearchQuery("");
-      setIsNetworkLoading(true);
-    })();
-  };
-
-  // 外部クリックを検出してフォーカスを解除
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      const searchContainer = document.getElementById("search-container");
-      if (searchContainer && !searchContainer.contains(event.target as Node)) {
-        setIsFocused(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
 
   if (isNetworkLoading || !steamAllData) {
     return <Loading />;
@@ -264,104 +178,15 @@ const Network = (props: Props) => {
       <div className="flex-1 relative bg-gray-900 overflow-hidden z-10">
         <HomeHeader />
 
-        {/* 検索バー */}
-        <div
-          id="search-container"
-          className={`absolute top-4 left-0 z-30 py-2 rounded-lg backdrop-filter backdrop-blur-sm transition-all duration-300 ${
-            isSidebarOpen ? "ml-72" : "ml-8" // Sidebarの幅に応じてマージンを変更
-          }`}
-        >
-          {/* 検索フォーム */}
-          <div className="flex flex-col">
-            <div className="relative flex items-center">
-              <input
-                type="text"
-                placeholder="ゲームタイトルを検索"
-                value={searchQuery}
-                onChange={(e) => setSearchQuery(e.target.value)}
-                onFocus={() => setIsFocused(true)}
-                className={`w-full p-2 pl-8 text-black border-2 border-gray-200 focus:outline-none transition duration-300 ease-in-out ${
-                  isFocused &&
-                  searchQuery !== "" &&
-                  filteredSteamList.length > 0
-                    ? "rounded-t-lg"
-                    : "rounded-lg"
-                }`}
-              />
-              <SearchIcon className="absolute left-2 text-gray-500" />
-              {searchQuery && (
-                <button
-                  onClick={() => setSearchQuery("")}
-                  className="absolute right-2 text-gray-500 hover:text-gray-700 focus:outline-none"
-                  aria-label="クリア検索"
-                >
-                  ×
-                </button>
-              )}
-            </div>
-
-            {/* メッセージ表示エリア */}
-            {searchQuery !== "" && filteredSteamList.length === 0 && (
-              <p className="text-gray-300 mt-1">
-                該当するゲームが見つかりません。別のゲームをお探しください。
-              </p>
-            )}
-
-            {/* ゲーム追加候補表示 */}
-            {isFocused &&
-              searchQuery !== "" &&
-              filteredSteamList.length > 0 && (
-                <div className="bg-white text-black rounded-b-lg max-h-60 overflow-y-auto cursor-pointer">
-                  {filteredSteamList.map((game) => (
-                    <div key={"filteredSteamList" + game.steamGameId}>
-                      {typeof game.index === "number" ? (
-                        <div
-                          className="cursor-pointer flex items-center hover:bg-gray-400 px-2"
-                          onClick={() => {
-                            if (game.index !== undefined) {
-                              setSelectedIndex(game.index);
-                              setSearchQuery(game.title);
-                              setIsFocused(false);
-                              setIsGameSearchOpen(true);
-                            }
-                          }}
-                        >
-                          <div
-                            className="text-center"
-                            style={{
-                              minWidth: "40px",
-                            }}
-                          >
-                            {game.index + 1}位
-                          </div>
-                          <div className="p-2">{game.title}</div>
-                        </div>
-                      ) : (
-                        <div
-                          className="flex items-center hover:bg-gray-400 px-2"
-                          onClick={() => {
-                            handleSearchClick(game.steamGameId);
-                            setIsFocused(false);
-                          }}
-                        >
-                          <div
-                            style={{
-                              minWidth: "40px",
-                              display: "flex",
-                              justifyContent: "center",
-                            }}
-                          >
-                            <PlaylistAddIcon className="cursor-pointer hover:bg-gray-600 rounded" />
-                          </div>
-                          <div className="p-2">{game.title}</div>
-                        </div>
-                      )}
-                    </div>
-                  ))}
-                </div>
-              )}
-          </div>
-        </div>
+        <SearchGames
+          setSelectedIndex={setSelectedIndex}
+          userAddedGames={userAddedGames}
+          setUserAddedGames={setUserAddedGames}
+          nodes={nodes}
+          setIsNetworkLoading={setIsNetworkLoading}
+          steamListData={steamListData}
+          openPanel={openPanel}
+        />
 
         {/* ゲーム詳細表示 */}
         {selectedIndex !== -1 && nodes[selectedIndex] && isGameSearchOpen && (
